argPass 0.5.1

+++ userprog/process.c
@@ -633,10 +633,10 @@ setup_stack (struct intr_frame *if_, const int argc, char **argv) {
 		memset (esp, 0, sizeof (void*));
 		/* Set RDI. */
 		if_->R.rdi = (uint64_t)argc;
-		
+
 		for (uint8_t *i = (uint8_t*)(USER_STACK -1); i >= esp; i--)
 			printf("i: 0x%x, val: 0x%x, char: %c\n", i, *i, *i);
-		printf("esp: %x, rdi: %d, rsi: %x\n", esp, (int)if_->R.rdi, if_->R.rsi);
+		printf("esp: %x, rdi: %d, rsi: %x\n", esp, (int)if_->R.rdi, (uint64_t)if_->R.rsi);
 		ASSERT(0);
 	}
 	////////////////////////////////////////////////////////////////////////////////////////////////////////
